Refactor job caching to use project-specific cache paths
