+++ cluv/utils.py
@@ -1,6 +1,7 @@
 import os
 import socket
 import sys
+from pathlib import Path
 
 import rich.console
 
@@ -14,3 +15,15 @@ def current_cluster() -> str | None:
     if "CC_CLUSTER" in os.environ:
         return os.environ["CC_CLUSTER"]
     return None
+
+
+def find_pyproject(start: Path | None = None) -> Path:
+    current = (start or Path.cwd()).resolve()
+    for folder in (current, *current.parents):
+        candidate = folder / "pyproject.toml"
+        if candidate.is_file():
+            return candidate
+    raise RuntimeError(
+        f"Could not find pyproject.toml starting from {current}!\n"
+        f"Cluv can only be used within a project managed with uv."
+    )
